drone_ops_agent/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
openai
